A functional prototype of Janrain social login integration.  Currently it's displayed on the Fundraising Campaign template, but it can be placed anywhere you can drop a div tag with an id.
Creates a regular plone user from the callback data after social login and uses the email address as the key.  So, for example, this works with my personal gmail account and my Facebook account as the same user to Plone because the both profiles have the same email address.  The sync with Salesforce still works too.

To do:

- Integrate into login and reg forms
- Integrate social identities with Social Contacts in SF