Fixed the following issues:     - Implement hide amount field     - Show confirmation of honorary on receipt     - Show live preview of email while editing
I had to change the way lookup of donations for building receipts was handled from a cached view method to a cached instance method on FundraisingCampaignPage.  I couldn't figure out how to expire an item from the plone.memoize view cache and this seems cleaner anyhow.  The HonoraryMemorialView now clears the cache after it updates the opportunity in salesforce.