New login/register user experience implementation.
This was triggered by the use case where a person donates and the system creates a user account in the backend for them.  Then the same person returns and attempts to register for a site account using the same email address.  They are then told that they already have an account though they never knew about the background user account that their donation created.  The background account is desired functionality b/c it caches their contact info so if they later come to claim the account, we don't have to prompt them for everything again.

The solution was to change the login/register flow for site accounts.  The Janrain social sign flow did not change.  The site account login/register process is now:

1. Prompt the user for their email

2. Send to /email-login-redirect which determines which category the user falls into:
   - registered: they have previously registered and set their own password
   - social: they have previously registered via social sign in but never set their own password
   - donor_only: they have an account created in the background from a donation and have not set a password
   - None: there is no existing account for the email address

3. User is redirected to...
    - Login form if "registered"
    - Set Password form if "social" or "donor_only"
    - Create User form if None
    * Each redirect respects came_from, prefills the passed email, and ultimately logs the user in automatically at the appropriate point in time

Implemenation details:
    - Added support via javascript on pluggable login page for ?tab=X where X is the tab index.  If passed, the tab specified will be selected instead of the first tab
    - Added fields for registered and social_signin to Person to track login sources